requirements.txt
fastapi
uvicorn
spotipy
requests
pillow